requirements.txt
streamlit==1.35.0
tensorflow==2.15.0
pillow==10.0.0
matplotlib==3.8.0
numpy==1.26.0
opencv-python==4.8.0
